requirements.txt
Django==1.9.7
djangorestframework
MySQL-python
django-allauth
stripe
django-tinymce
